BETA_VERSION 1.0 Published
The first upload of my website and it has nothing in it and it has a lot of bugs. Feel Free to check whats in it and you can make change to it and i dont actually gives a shit. 
Looking forward to when the website is uploaded to the internet and for everyone to take a look!